move environment model in model folder
