+++ src/app/tokens/api/apis.token.ts
@@ -1,6 +1,6 @@
 import { InjectionToken } from '@angular/core';
 import { ApiConfiguration } from './api.configuration';
-import { Api } from '../../../environments/environment.model';
+import { Api } from '../../../environments';
 
 export const APIS_TOKENS: InjectionToken<Record<Api, ApiConfiguration>> = new InjectionToken<Record<Api, ApiConfiguration>>(
   'apis.configuration'
