+++ src/environments/index.ts
@@ -1,2 +1,3 @@
 export * from './environment';
 export * from './enum/environement';
+export * from './models/environment.model';
