+++ src/environments/models/environment.model.ts
@@ -1,5 +1,5 @@
-import { EnvironmentType } from './enum/environement';
-import { ApiConfiguration } from '../app/tokens';
+import { EnvironmentType } from '../enum/environement';
+import { ApiConfiguration } from '../../app/tokens';
 
 export enum Api {
   ConseillerNumerique = '@conseillerNumerique',
